logger: log dropout events
